+++ src/mainwindow.cpp
@@ -1,6 +1,7 @@
 #include "mainwindow.h"
 #include "ui_mainwindow.h"
 #include "scene.h"
+#include "menu.h"
 
 MainWindow::MainWindow(QWidget *parent)
     : QMainWindow(parent)
@@ -10,7 +11,19 @@ MainWindow::MainWindow(QWidget *parent)
 
     resize(750, 750);
     Scene *scene = initScene();
-    Q_UNUSED(scene); //! just unused but prepared to be used
+
+    QGraphicsView *view = ui->graphicsView;
+    view->setScene(scene);
+    view->setFocusPolicy(Qt::StrongFocus);
+
+    connect(ui->Right_menu, &Menu::start, scene, [scene, view]() {
+        scene->start();
+        view->setFocus();
+    });
+    connect(ui->Right_menu, &Menu::stop, scene, [scene, view]() {
+        scene->stop();
+        view->setFocus();
+    });
 }
 
 MainWindow::~MainWindow()
@@ -29,7 +42,7 @@ Scene *MainWindow::initScene()
     Scene *scene = new Scene(ui->graphicsView);
     ui->graphicsView->setScene(scene);
 
-    ui->graphicsView->setSceneRect(0, 0, 400, 700);
+    ui->graphicsView->setSceneRect(0, 0, 400, 800);
     ui->graphicsView->setRenderHint(QPainter::Antialiasing);
     ui->graphicsView->setResizeAnchor(QGraphicsView::NoAnchor);
     return scene;