+++ CHANGELOG.md
@@ -6,3 +6,10 @@ All notable changes to this project will be documented in this file.
 
 * Started this project
 * Set up first window
+
+## 30.7.2024
+
+* User can spawn blocks, that will just fall down
+* User can move blocks to the left and right
+* Fixed size of graphics scene
+* Only last spawned block can be controlled